@@ -5,7 +5,8 @@
   "main": "package.json",
   "author": "Sanity <hello@sanity.io>",
   "scripts": {
-    "dev": "sanity start",
+    "start": "sanity start",
+    "dev": "sanity dev",
     "format": "prettier-eslint --write \"**/*.js\" \"!node_modules/**\"",
     "build": "sanity build",
     "graphql-deploy": "sanity graphql deploy --playground",
@@ -13,22 +14,21 @@
     "test": "sanity check"
   },
   "dependencies": {
-    "@sanity/base": "^2.1.3",
     "@sanity/cli": "^2.1.0",
     "@sanity/components": "^2.1.0",
-    "@sanity/core": "^2.1.2",
     "@sanity/dashboard": "^2.1.0",
-    "@sanity/default-layout": "^2.1.3",
-    "@sanity/default-login": "^2.1.0",
-    "@sanity/desk-tool": "^2.1.3",
     "@sanity/studio-hints": "^2.1.0",
+    "@sanity/vision": "^3.21.1",
     "date-fns": "^1.30.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^3.11.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.2.0",
+    "sanity": "^3.21.1",
     "sanity-plugin-dashboard-widget-document-list": "^0.0.11",
-    "sanity-plugin-dashboard-widget-netlify": "^1.0.1"
+    "sanity-plugin-dashboard-widget-netlify": "^1.0.1",
+    "styled-components": "^6.1.1",
+    "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
