@@ -1,41 +1,36 @@
 export default {
-  name: 'siteSettings',
-  type: 'document',
-  title: 'Site Settings',
-  __experimental_actions: [
-    // 'create',
-    'update',
-    // 'delete',
-    'publish'
-  ],
+  name: "siteSettings",
+  type: "document",
+  title: "Site Settings",
+  __experimental_actions: ["update", /* 'create', 'delete', */ "publish"],
   fields: [
     {
-      name: 'title',
-      type: 'string',
-      title: 'Title'
+      name: "title",
+      type: "string",
+      title: "Title",
     },
     {
-      name: 'description',
-      type: 'text',
-      title: 'Description',
-      description: 'Describe your portfolio for search engines and social media.'
+      name: "description",
+      type: "text",
+      title: "Description",
+      description: "Describe your blog for search engines and social media.",
     },
     {
-      name: 'keywords',
-      type: 'array',
-      title: 'Keywords',
-      description: 'Add keywords that describes your portfolio.',
-      of: [{type: 'string'}],
+      name: "keywords",
+      type: "array",
+      title: "Keywords",
+      description: "Add keywords that describes your blog.",
+      of: [{ type: "string" }],
       options: {
-        layout: 'tags'
-      }
+        layout: "tags",
+      },
     },
     {
-      name: 'author',
-      type: 'reference',
-      description: 'Publish an author and set a reference to them here.',
-      title: 'Author',
-      to: [{type: 'person'}]
-    }
-  ]
-}
+      name: "author",
+      type: "reference",
+      description: "Publish an author and set a reference to them here.",
+      title: "Author",
+      to: [{ type: "author" }],
+    },
+  ],
+};
