@@ -1,34 +1,20 @@
 export default {
-  name: 'bioPortableText',
-  type: 'array',
-  title: 'Excerpt',
+  name: "bioPortableText",
+  type: "array",
+  title: "Excerpt",
   of: [
     {
-      title: 'Block',
-      type: 'block',
-      styles: [{title: 'Normal', value: 'normal'}],
+      type: "block",
+      title: "Block",
+      styles: [{ title: "Normal", value: "normal" }],
       lists: [],
       marks: {
         decorators: [
-          {title: 'Strong', value: 'strong'},
-          {title: 'Emphasis', value: 'em'},
-          {title: 'Code', value: 'code'}
+          { title: "Strong", value: "strong" },
+          { title: "Emphasis", value: "em" },
+          { title: "Code", value: "code" },
         ],
-        annotations: [
-          {
-            name: 'link',
-            type: 'object',
-            title: 'URL',
-            fields: [
-              {
-                title: 'URL',
-                name: 'href',
-                type: 'url'
-              }
-            ]
-          }
-        ]
-      }
-    }
-  ]
-}
+      },
+    },
+  ],
+};
